+++ dragonchain/broadcast_processor/broadcast_processor.py
@@ -196,7 +196,7 @@ async def send_notification_verification(
     await broadcast_functions.remove_notification_verification_for_broadcast_async(redis_list_value)
 
 
-async def process_blocks_for_broadcast(session: aiohttp.ClientSession) -> None:
+async def process_blocks_for_broadcast(session: aiohttp.ClientSession) -> None:  # noqa: C901
     """Main function of the broadcast processor
 
     Retrieves blocks that need to be processed, gets matchmaking claims for blocks,
@@ -219,10 +219,17 @@ async def process_blocks_for_broadcast(session: aiohttp.ClientSession) -> None:
             _log.warning("[BROADCAST PROCESSOR] Out of funds! Will not broadcast anything for 30 minutes")
             await asyncio.sleep(1800)  # Sleep for 30 minutes if insufficient funds
             break
-        except exceptions.NotFound:
+        except exceptions.UnableToUpdate:
             _log.warning("Matchmaking does not have enough matches to create a claim check")
             # Schedule this block for 5 minutes later, so we don't spam matchmaking every second if there aren't matches available
             await broadcast_functions.schedule_block_for_broadcast_async(block_id, int(time.time()) + 300)
+            continue
+        except exceptions.NotFound:
+            _log.warning(
+                f"Matchmaking does not have record of claim for block {block_id}."
+                "Presumably closed. Saving to unfinished claim and removing from broadcast system."
+            )
+            await broadcast_functions.save_unfinished_claim(block_id)
             continue
         claim_chains = chain_id_set_from_matchmaking_claim(claim, current_level)
         if score == 0:
@@ -247,21 +254,28 @@ async def process_blocks_for_broadcast(session: aiohttp.ClientSession) -> None:
                     _log.info(f"[BROADCAST PROCESSOR] Chain {chain} didn't respond to broadcast in time. Fetching new chain")
                     try:
                         claim = matchmaking.overwrite_no_response_node(block_id, current_level, chain)
-                    except exceptions.NotFound:
+                    except exceptions.UnableToUpdate:
                         _log.warning(f"Matchmaking does not have enough matches to update this claim check with new chains for level {current_level}")
+                        # Schedule for 5 minutes later, so we don't spam matchmaking every second if there aren't matches
+                        await broadcast_functions.schedule_block_for_broadcast_async(block_id, int(time.time()) + 300)
+                        claim = None
+                        break
+                    except exceptions.NotFound:
+                        _log.warning(
+                            f"Matchmaking does not have record of claim for block {block_id}."
+                            "Presumably closed. Saving to unfinished claim and removing from broadcast system."
+                        )
+                        await broadcast_functions.save_unfinished_claim(block_id)
                         claim = None
                         break
                 # Can't continue processing this block if the claim wasn't updated
                 if claim is None:
-                    # Schedule for 5 minutes later, so we don't spam matchmaking every second if there aren't matches
-                    await broadcast_functions.schedule_block_for_broadcast_async(block_id, int(time.time()) + 300)
                     continue
                 new_claim_chains = chain_id_set_from_matchmaking_claim(claim, current_level)
                 # Make requests for all the new chains
                 futures = make_broadcast_futures(session, block_id, current_level, new_claim_chains.difference(current_verifications))
-                if (
-                    futures is None
-                ):  # This occurs when make_broadcast_futures failed to create the broadcast dto (we need to process this block later)
+                if futures is None:
+                    # This occurs when make_broadcast_futures failed to create the broadcast dto (we need to process this block later)
                     continue
                 request_futures.update(futures)
                 # Schedule this block to be re-checked after BROADCAST_RECEIPT_WAIT_TIME more seconds have passed