+++ dragonchain/transaction_processor/level_1_actions.py
@@ -46,8 +46,8 @@ def execute() -> None:
     if BROADCAST:
         try:
             matchmaking.renew_registration_if_necessary()
-        except exceptions.MatchmakingError:
-            _log.warning("Could not register with matchmaking! Is your dragon net configuration valid?")
+        except (exceptions.MatchmakingError, exceptions.MatchmakingRetryableError):
+            _log.warning("Could not register with matchmaking! Is your Dragon Net configuration valid?")
     t0 = time.time()
 
     # Pop off of queue